HIVE-25607: Mask totalSize table property in Iceberg q-tests (Marton Bod, reviewed by Peter Vary and Adam Szita)
The totalSize table property can change whenever the file format version changes, therefore potentially causing the q tests to be flaky when issuing describe formatted commands. We should mask this and not test against the exact value.
