+++ iceberg/iceberg-handler/src/test/queries/positive/alter_part_table_to_iceberg.q
@@ -1,3 +1,6 @@
+-- Mask the totalSize value as it can change whenever the file format (e.g. ORC) version changes
+--! qt:replace:/(\s+totalSize\s+)[\S+](\s+)/$1#Masked#$2/
+
 set hive.vectorized.execution.enabled=false;
 
 drop table if exists tbl_orc;
