+++ iceberg/iceberg-handler/src/test/queries/positive/describe_iceberg_metadata_tables.q
@@ -1,9 +1,3 @@
-set hive.vectorized.execution.enabled = false;
-set tez.mrreader.config.update.properties=hive.io.file.readcolumn.names,hive.io.file.readcolumn.ids;
-set hive.query.results.cache.enabled=false;
-set hive.fetch.task.conversion=none;
-set hive.cbo.enable=true;
-
 drop table if exists ice_meta_desc;
 create external table ice_meta_desc (id int, value string) stored by iceberg stored as orc;
 insert into ice_meta_desc values (1, 'one'),(2,'two'),(3,'three'),(4,'four'),(5,'five');
