+++ iceberg/iceberg-handler/src/test/queries/positive/truncate_force_iceberg_table.q
@@ -1,4 +1,6 @@
 -- SORT_QUERY_RESULTS
+-- Mask the totalSize value as it can change whenever the file format (e.g. ORC) version changes
+--! qt:replace:/(\s+totalSize\s+)[\S+](\s+)/$1#Masked#$2/
 
 set hive.vectorized.execution.enabled=false;
