+++ iceberg/iceberg-handler/src/test/results/positive/truncate_force_iceberg_table.q.out
@@ -165,7 +165,7 @@ Table Parameters:
 	serialization.format	1                   
 	storage_handler     	org.apache.iceberg.mr.hive.HiveIcebergStorageHandler
 	table_type          	ICEBERG             
-	totalSize           	0                   
+	totalSize           	#Masked#                   
 #### A masked pattern was here ####
 	write.format.default	parquet
