+++ itests/qtest-iceberg/src/test/java/org/apache/hadoop/hive/cli/TestIcebergLlapLocalCliDriver.java
@@ -15,11 +15,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.apache.hadoop.hive.cli;
 
 import java.io.File;
 import java.util.List;
+
 import org.apache.hadoop.hive.cli.control.CliAdapter;
 import org.apache.hadoop.hive.cli.control.CliConfigs;
+
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
@@ -29,9 +32,9 @@ import org.junit.runners.Parameterized;
 import org.junit.runners.Parameterized.Parameters;
 
 @RunWith(Parameterized.class)
-public class TestIcebergCliDriver {
+public class TestIcebergLlapLocalCliDriver {
 
-  static CliAdapter adapter = new CliConfigs.IcebergCliConfig().getCliAdapter();
+  static CliAdapter adapter = new CliConfigs.IcebergLlapLocalCliConfig().getCliAdapter();
 
   @Parameters(name ="{0}")
   public static List<Object[]> getParameters() throws Exception {
@@ -47,7 +50,7 @@ public class TestIcebergCliDriver {
   private String name;
   private File qfile;
 
-  public TestIcebergCliDriver(String name, File qfile) {
+  public TestIcebergLlapLocalCliDriver(String name, File qfile) {
     this.name = name;
     this.qfile = qfile;
   }
